+++ Terminal-iOS/MyStudy/MyStudyViewController.swift
@@ -1,0 +1,18 @@
+//
+//  MyStudyViewController.swift
+//  Terminal-iOS
+//
+//  Created by 정재인 on 2020/08/31.
+//  Copyright © 2020 정재인. All rights reserved.
+//
+
+import UIKit
+
+class MyStudyViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .blue
+
+    }
+}
