--- Terminal-iOS/MyStudyViewController.swift
@@ -1,30 +1,0 @@
-//
-//  MyStudyViewController.swift
-//  Terminal-iOS
-//
-//  Created by 정재인 on 2020/08/31.
-//  Copyright © 2020 정재인. All rights reserved.
-//
-
-import UIKit
-
-class MyStudyViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .blue
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
