+++ Terminal-iOS/Study/CreateStudyViewController.swift
@@ -1,0 +1,18 @@
+//
+//  CreateStudyViewController.swift
+//  Terminal-iOS
+//
+//  Created by once on 2020/08/31.
+//  Copyright © 2020 정재인. All rights reserved.
+//
+
+import UIKit
+
+class CreateStudyViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .darkGray
+        // Do any additional setup after loading the view.
+    }
+}
