+++ Terminal-iOS/Study/StudyViewController.swift
@@ -1,0 +1,23 @@
+//
+//  StudyViewController.swift
+//  Terminal-iOS
+//
+//  Created by 정재인 on 2020/08/31.
+//  Copyright © 2020 정재인. All rights reserved.
+//
+
+import UIKit
+
+class StudyViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        self.navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(createStudy))
+    }
+    
+    @objc func createStudy() {
+        let view = CreateStudyViewController()
+        navigationController?.pushViewController(view, animated: true)
+    }
+}
